Moves get_parser_install_dir into configs out of utils.
Previously `get_parser_install_dir` and `get_parser_info_dir` were both in
utils. This is because they only took into the package dir and the site
dir.
Now there is a configuration option, `parser_install_dir` that takes
precedence over the package and site directories. This is parsed in with
``` lua
configs.setup(
{
    parser_install_dir = "some/path",
})
```

As such the `get_parser_install_dir` needs to take this option in. One
was is to leave it in usils and parse it in:
```lua
utils.get_parser_install_dir("parser", configs.get_parser_install_dir_option())
```
But that will always require access to configs, or will work
inconsistently. So it was easier to move the whole setting into configs:
```lua
configs.get_parser_install_dir("parser")
```
